# internal/app/cloudgontroller/api/swagger/swagger.yaml
basePath: /api/v3
definitions:
  controllers.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  presenter.BuildpackResponse:
    properties:
      created_at:
        type: string
      enabled:
        type: boolean
      filename:
        type: string
      guid:
        type: string
      links:
        properties:
          self:
            properties:
              href:
                type: string
            type: object
          upload:
            properties:
              href:
                type: string
              method:
                type: string
            type: object
        type: object
      locked:
        type: boolean
      metadata:
        $ref: '#/definitions/presenter.Metadata'
      name:
        type: string
      position:
        type: integer
      stack:
        type: string
      state:
        type: string
      updated_at:
        type: string
    type: object
  presenter.Metadata:
    properties:
      annotations:
        type: object
      labels:
        type: object
    type: object
host: localhost:8080
info:
  contact: {}
  description: CAPI V3 Compatible API with blazing fast backend.
  title: CloudGontroller API
  version: 3.0.0
paths:
  /buildpacks:
    get:
      consumes:
      - application/json
      description: Retrieve all buildpacks the user has access to.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenter.BuildpackResponse'
        "400":
          description: Bad Request
          schema:
            items:
              type: object
            type: array
        "404":
          description: Not Found
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.HTTPError'
      summary: Buildpacks List buildpacks
      tags:
      - Buildpacks
  /buildpacks/{guid}:
    get:
      description: Retrieve all buildpacks the user has access to.
      parameters:
      - description: Buildpack GUID
        in: path
        name: guid
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenter.BuildpackResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.HTTPError'
        "404":
          description: Not Found
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.HTTPError'
      summary: Show a buildpack
      tags:
      - Buildpacks
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"
x-extension-openapi:
  example: value on a json format